feat(form): add reatomFieldSet docs, refactor reatomFieldSet
