+++ packages/form/src/index.ts
@@ -1,2 +1,3 @@
 export * from './reatomField'
 export * from './reatomForm'
+export * from './reatomFieldSet'
